requirements.txt
# Core uAgents framework (matching official Fetch.ai examples)
openai>=1.0.0
hyperon>=0.2.6  # ✅ SingularityNET's MeTTa for knowledge graphs (mailbox agents only!)
uagents>=0.22.5
uagents-core>=0.3.5
python-dotenv>=1.0.0

# DeFi and web3 support
web3>=6.15.0

# Additional dependencies
requests>=2.31.0
aiohttp>=3.9.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.8.0
pandas>=2.1.4
numpy>=2.0.2
python-dateutil>=2.8.2
colorama>=0.4.6
rich>=13.7.0
